Rodriguez and Mullin relationship for mass / size
Jessica pointed me to a paper from 1986 that provides an approximation for
biomass given phytoplankton size (I didn't see them explicitly say ESD but it
seems like a reasonable assumption). It's a log-linear relationship, so I moved
terms around and use it to compute ESD for a given biomass rather than assuming
a linear relationship.